requirements.txt
dependency-injector
flask
bootstrap-flask
pygithub
pyyaml
pytest-flask
pytest-cov